split packet_map into server and client halves, move is_alive to client, move connection code to connections.rs, player list in status
